requirements.txt
MetaTrader5
pandas
pyTelegramBotAPI
python-dotenv
pytz
